feat: CytoscapeME directory renamed to ModelVisualization
