+++ src/services/global/LiveModel/ExportLM.ts
@@ -4,7 +4,7 @@ import type {
   ModelStats,
   Variable,
 } from '../../../types';
-import CytoscapeME from '../../model-editor/CytoscapeME/CytoscapeME';
+import CytoscapeME from '../../model-editor/ModelVisualization/CytoscapeME';
 import type { LiveModelClass } from './LiveModel';
 import useVariablesStore from '../../../stores/LiveModel/useVariablesStore';
 import useUpdateFunctionsStore from '../../../stores/LiveModel/useUpdateFunctionsStore';
