+++ src/services/global/LiveModel/VariablesLM.ts
@@ -1,7 +1,7 @@
 import useUpdateFunctionsStore from '../../../stores/LiveModel/useUpdateFunctionsStore';
 import useVariablesStore from '../../../stores/LiveModel/useVariablesStore';
 import type { ControlInfo, Position, Variable } from '../../../types';
-import CytoscapeME from '../../model-editor/CytoscapeME/CytoscapeME';
+import CytoscapeME from '../../model-editor/ModelVisualization/CytoscapeME';
 import type { LiveModelClass } from './LiveModel';
 import useRegulationsStore from '../../../stores/LiveModel/useRegulationsStore';
 import useControlStore from '../../../stores/LiveModel/useControlStore';
